fix norm with refactoring handle_keydown
