+++ Makefile
@@ -11,6 +11,7 @@ SRC				= $(SRC_DIR)/main.c \
 					$(SRC_DIR)/extract_map_info.c \
 					$(SRC_DIR)/check_map.c \
 					$(SRC_DIR)/plot_line.c \
+					$(SRC_DIR)/keydown_events.c \
 					$(SRC_DIR)/convert_points_to_isometric.c \
 					$(SRC_DIR)/render_image.c \
 					$(SRC_DIR)/handle_events.c \
