+++ src/extract_map_info.c
@@ -6,7 +6,7 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/27 17:18:59 by reasuke           #+#    #+#             */
-/*   Updated: 2023/12/13 19:00:32 by reasuke          ###   ########.fr       */
+/*   Updated: 2023/12/15 19:14:52 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -83,8 +83,6 @@ static void	_set_points(t_map *map, char **line_array)
 		y++;
 	}
 	y = 0;
-	map->min_val = INT_MAX;
-	map->max_val = INT_MIN;
 	while (y < map->height)
 	{
 		x = 0;
@@ -95,8 +93,6 @@ static void	_set_points(t_map *map, char **line_array)
 				map->points[y][x].color = 0xAABBC;
 			else if (map->points[y][x].color == NOT_SPECIFIED)
 				map->points[y][x].color = 0xDDDDDD;
-			ft_chmin(&map->min_val, map->points[y][x].z);
-			ft_chmax(&map->max_val, map->points[y][x].z);
 			x++;
 		}
 		y++;
